+++ src/uniprotptmpy/__init__.py
@@ -4,6 +4,7 @@ from uniprotptmpy._download import download
 from uniprotptmpy._ptmlist_writer import write_ptmlist
 from uniprotptmpy._tabular import write_tsv
 from uniprotptmpy.database import PtmDatabase
+from uniprotptmpy.errors import UniprotPtmError, UniprotPtmParseError
 from uniprotptmpy.models import CrossReference, FeatureType, PtmEntry, TaxonomicRange
 from uniprotptmpy.parser import load, parse_ptm_list
 
@@ -15,6 +16,8 @@ __all__ = [
     "FeatureType",
     "PtmEntry",
     "TaxonomicRange",
+    "UniprotPtmError",
+    "UniprotPtmParseError",
     "PtmDatabase",
     "download",
     "load",